# Frontend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Enterprise Project Management System",
        "description": "Production-ready FastAPI backend with workflow automation and AI integration",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/workflow/idea/{idea_id}": {
            "get": {
                "tags": [
                    "Workflows"
                ],
                "summary": "Get Idea Workflow State",
                "description": "Get current workflow state for an idea\nReturns current step, allowed actions, responsible roles/teams",
                "operationId": "get_idea_workflow_state_api_workflow_idea__idea_id__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "idea_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Idea Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WorkflowStateResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/workflow/idea/{idea_id}/action": {
            "post": {
                "tags": [
                    "Workflows"
                ],
                "summary": "Perform Idea Workflow Action",
                "description": "Perform a workflow action on an idea\nValidates permissions, updates idea status, workflow instance, and logs history",
                "operationId": "perform_idea_workflow_action_api_workflow_idea__idea_id__action_post",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "idea_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Idea Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WorkflowActionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WorkflowActionResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/workflow/idea/{idea_id}/history": {
            "get": {
                "tags": [
                    "Workflows"
                ],
                "summary": "Get Idea Workflow History",
                "description": "Get workflow history for an idea",
                "operationId": "get_idea_workflow_history_api_workflow_idea__idea_id__history_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "idea_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Idea Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/WorkflowHistoryResponse"
                                    },
                                    "title": "Response Get Idea Workflow History Api Workflow Idea  Idea Id  History Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/workflow/available-actions/{entity_type}/{entity_id}": {
            "get": {
                "tags": [
                    "Workflows"
                ],
                "summary": "Get Available Workflow Actions",
                "description": "Get available workflow actions for a user on a specific entity",
                "operationId": "get_available_workflow_actions_api_workflow_available_actions__entity_type___entity_id__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "entity_type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Entity Type"
                        }
                    },
                    {
                        "name": "entity_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Entity Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/workflow/my-tasks": {
            "get": {
                "tags": [
                    "Workflows"
                ],
                "summary": "Get Workflow Tasks For User",
                "description": "Get workflow tasks assigned to the current user",
                "operationId": "get_workflow_tasks_for_user_api_workflow_my_tasks_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
            "WorkflowActionInfo": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "next_step": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Next Step"
                    },
                    "is_terminal": {
                        "type": "boolean",
                        "title": "Is Terminal",
                        "default": false
                    }
                },
                "type": "object",
                "required": [
                    "name"
                ],
                "title": "WorkflowActionInfo",
                "description": "Workflow action information"
            },
            "WorkflowActionRequest": {
                "properties": {
                    "action": {
                        "type": "string",
                        "title": "Action"
                    },
                    "notes": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Notes"
                    }
                },
                "type": "object",
                "required": [
                    "action"
                ],
                "title": "WorkflowActionRequest",
                "description": "Workflow action request"
            },
            "WorkflowActionResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "title": "Success"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message"
                    },
                    "new_status": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "New Status"
                    },
                    "new_step": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "New Step"
                    },
                    "is_terminal": {
                        "type": "boolean",
                        "title": "Is Terminal",
                        "default": false
                    }
                },
                "type": "object",
                "required": [
                    "success",
                    "message"
                ],
                "title": "WorkflowActionResponse",
                "description": "Workflow action response"
            },
            "WorkflowHistoryResponse": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "step_name": {
                        "type": "string",
                        "title": "Step Name"
                    },
                    "action": {
                        "type": "string",
                        "title": "Action"
                    },
                    "actor_name": {
                        "type": "string",
                        "title": "Actor Name"
                    },
                    "notes": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Notes"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "step_name",
                    "action",
                    "actor_name",
                    "created_at"
                ],
                "title": "WorkflowHistoryResponse",
                "description": "Workflow history response"
            },
            "WorkflowResponsibleInfo": {
                "properties": {
                    "type": {
                        "type": "string",
                        "title": "Type"
                    },
                    "role": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Role"
                    },
                    "department": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Department"
                    }
                },
                "type": "object",
                "required": [
                    "type"
                ],
                "title": "WorkflowResponsibleInfo",
                "description": "Workflow responsible information"
            },
            "WorkflowStateResponse": {
                "properties": {
                    "entity_id": {
                        "type": "integer",
                        "title": "Entity Id"
                    },
                    "entity_type": {
                        "type": "string",
                        "title": "Entity Type"
                    },
                    "current_step": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/WorkflowStepInfo"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "available_actions": {
                        "items": {
                            "$ref": "#/components/schemas/WorkflowActionInfo"
                        },
                        "type": "array",
                        "title": "Available Actions",
                        "default": []
                    },
                    "responsible": {
                        "items": {
                            "$ref": "#/components/schemas/WorkflowResponsibleInfo"
                        },
                        "type": "array",
                        "title": "Responsible",
                        "default": []
                    },
                    "can_user_act": {
                        "type": "boolean",
                        "title": "Can User Act",
                        "default": false
                    }
                },
                "type": "object",
                "required": [
                    "entity_id",
                    "entity_type",
                    "status"
                ],
                "title": "WorkflowStateResponse",
                "description": "Workflow state response"
            },
            "WorkflowStepInfo": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "status": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Status"
                    },
                    "is_terminal": {
                        "type": "boolean",
                        "title": "Is Terminal",
                        "default": false
                    },
                    "available_actions": {
                        "items": {
                            "$ref": "#/components/schemas/WorkflowActionInfo"
                        },
                        "type": "array",
                        "title": "Available Actions",
                        "default": []
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "name"
                ],
                "title": "WorkflowStepInfo",
                "description": "Workflow step information"
            }
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}